Add e2e setup
