+++ packages/e2e-tests/package.json
@@ -3,9 +3,12 @@
   "private": true,
   "description": "End-to-end tests for aragonCLI",
   "scripts": {
-    "test": "cross-env DEBUG=true ava",
-    "test:clean": "shx rm -rf .tmp && npm run reset-cache",
-    "test:watch": "cross-env DEBUG=true ava --watch",
+    "pretest": "ava src/**/*.pretest.js",
+    "posttest": "ava src/**/*.posttest.js",
+    "test": "ava",
+    "test:watch": "ava --watch",
+    "test:debug": "cross-env DEBUG=true npm run test",
+    "test:clean": "shx rm -rf '.tmp' && npm run reset-cache",
     "update-snapshots": "ava --update-snapshots",
     "reset-cache": "ava --reset-cache"
   },
@@ -21,10 +24,11 @@
     "@babel/preset-env": "^7.3.4",
     "@babel/register": "^7.0.0",
     "ava": "^2.0.0",
-    "cross-env": "^5.2.0",
     "create-aragon-app": "file:../create-aragon-app",
+    "cross-env": "^5.2.0",
     "execa": "^1.0.0",
     "fs-extra": "^8.0.1",
+    "kill-port": "^1.5.1",
     "node-fetch": "^2.3.0",
     "ps-tree": "^1.2.0"
   },
