+++ packages/e2e-tests/src/cli/devchain.posttest.js
@@ -1,0 +1,12 @@
+import test from 'ava'
+import killProcessOnPort from 'kill-port'
+
+test('should kill ganache', async t => {
+  t.plan(1)
+
+  // act
+  await killProcessOnPort('8545')
+
+  // assert
+  t.pass()
+})
