+++ packages/e2e-tests/src/cli/devchain.pretest.js.md
@@ -1,0 +1,35 @@
+# Snapshot report for `src/cli/devchain.pretest.js`
+
+The actual snapshot is saved in `devchain.pretest.js.snap`.
+
+Generated by [AVA](https://ava.li).
+
+## should spawn ganache
+
+> Snapshot 1
+
+    `> Could not find 'arapp.json'. Using the default configuration to connect to localhost.␊
+    Setting up a new chain from latest Aragon snapshot [started]␊
+    Setting up a new chain from latest Aragon snapshot [completed]␊
+    Starting a local chain from snapshot [started]␊
+    Local chain started at port 8545 [title changed]␊
+     i eth_accounts␊
+    Local chain started at port 8545 [completed]␊
+     i Here are some Ethereum accounts you can use.␊
+      The first one will be used for all the actions the CLI performs.␊
+      You can use your favorite Ethereum provider or wallet to import their private keys.␊
+      ␊
+    Address #1:  0xb4124cEB3451635DAcedd11767f004d8a28c6eE7 (this account is used to deploy DAOs, it has more permissions)␊
+    Private key: a8a54b2d8197bc0b19bb8a084031be71835580a01e70a45a13babd16c9bc1563␊
+    Address #2:  0x8401Eb5ff34cc943f096A32EF3d5113FEbE8D4Eb ␊
+    Private key: ce8e3bda3b44269c147747a373646393b1504bfcbb73fc9564f5d753d8116608␊
+     i The accounts were generated from the following mnemonic phrase:␊
+    explain tackle mirror kit van hammer degree position ginger unfair soup bonus␊
+    ␊
+     ⚠ The devchain was reset, some steps need to be done to prevent issues:␊
+        - Reset the application cache in Aragon Core by going to Settings > Troubleshooting.␊
+        - If using Metamask: switch to a different network, and then switch back to the 'Private Network' (this will clear the nonce cache and prevent errors when sending transactions)  ␊
+      ␊
+     i ENS instance deployed at 0x5f6f7e8cc7346a11ca2def8f827b7a0b612c56a1␊
+    ␊
+     i Devchain running: http://localhost:8545.`
