+++ packages/e2e-tests/src/cli/ipfs.pretest.js.md
@@ -1,6 +1,6 @@
-# Snapshot report for `src/cli/ipfs.test.js`
+# Snapshot report for `src/cli/ipfs.pretest.js`
 
-The actual snapshot is saved in `ipfs.test.js.snap`.
+The actual snapshot is saved in `ipfs.pretest.js.snap`.
 
 Generated by [AVA](https://ava.li).
